pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "orrery"
description = "A settlement-scale social simulation framework for games"
authors = [
  { name="Shi Johnson-Bey", email="shijbey@gmail.com" },
]
readme = "README.md"
dynamic = ["version"]
requires-python = ">=3.8"
keywords = [
    "social simulation",
    "orrery",
    "games",
    "simulation",
    "artificial intelligence",
    "agent-based modeling",
    "multiagent systems",
    "emergent narrative"
]
license = { file = "LICENSE.md" }
classifiers = [
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Games/Entertainment :: Simulation",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Artificial Life",
    "Topic :: Sociology",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "esper",
    "ordered-set",
    "numpy",
    "pyYAML",
    "pydantic",
    "tracery",
    "pandas"
]

[project.optional-dependencies]
testing = [
    "pytest",
    "pytest-cov"
]
development = [
    "isort",
    "black",
    "black[d]"
]
docs = [
    "sphinx"
]

[project.urls]
"Homepage" = "https://github.com/ShiJbey/orrery"
"Bug Tracker" = "https://github.com/ShiJbey/orrery/issues"
"Repository" = "https://github.com/ShiJBey/orrery.git"
"Changelog" = "https://github.com/ShiJbey/orrery/blob/main/CHANGELOG.md"

[project.scripts]
orrery = "orrery.__main__:main"

[tool.setuptools.dynamic]
version = {attr = "orrery.__version__"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
default_section = "THIRDPARTY"
known_first_party = "orrery"
src_paths = ["src/orrery", "test", "samples"]

[tool.pytest.ini_options]
minversion = "6.0"
testpaths = ["tests"]
